keep domensions as int for matrices and size_t for images
git-svn-id: file:///home/roberto/Git/svn@6508 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 7ae5ee03f6f7f766433cf1cd4255fc251954f9b2

+++ trunk/xmipp/libraries/data/matrix2d.h
@@ -374,7 +374,7 @@ public:
 
     /** Core allocate.
      */
-    void coreAllocate(int _mdimy, int _mdimx)
+    void coreAllocate( int _mdimy, int _mdimx)
     {
         if (_mdimy <= 0 ||_mdimx<=0)
         {
@@ -418,7 +418,7 @@ public:
         }
 
         T * new_mdata;
-        size_t YXdim=Ydim*Xdim;
+        int YXdim=Ydim*Xdim;
 
         try
         {
